"areej"

+++ Models/Chambre.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public int NumeroChambre { get; set; }
         public string Type { get; set; }
-        public double Prix { get; set; }
+        public float Prix { get; set; }
         public ICollection<Reservation> Reservations { get; set; }
